requirement.txt
fastembed 
chromadb==0.3.29 
pdfplumber 
langchain 
langchain-community 
langchain-text-splitters
boto3==1.34.30 
python-dotenv 
streamlit